Create translatePublicKey util
